Continue template decomposition: extract page templates
